requirements.txt
absl-py==2.1.0
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
astrapy==2.0.1
astunparse==1.6.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bitsandbytes==0.47.0
blinker==1.9.0
blis==1.2.0
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ChatterBot==1.2.2
chatterbot-corpus==1.2.2
chromadb==1.0.15
click==8.2.1
cloudpathlib==0.21.0
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
cryptography==45.0.5
cssselect2==0.8.0
ctransformers==0.2.27
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==4.0.0
deprecation==2.1.0
diff-match-patch==20241021
dill==0.3.8
distro==1.9.0
Django==5.1.7
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==23.5
django-import-export==4.1.1
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.10
duckdb==1.3.2
durationpy==0.10
ecdsa==0.19.1
email_validator==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
Flask-Mail==0.10.0
flatbuffers==25.2.10
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hf-xet==1.1.7
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jwt==1.3.1
keras==3.11.2
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.27
langchain-astradb==0.6.0
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.9
langchain-huggingface==0.3.1
langchain-openai==0.3.28
langchain-redis==0.2.3
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.4.11
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mathparse==0.1.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
motor==3.7.1
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.12
mypy_extensions==1.1.0
mysql-connector-python==8.0.33
mysqlclient==2.2.7
namex==0.0.8
narwhals==1.34.1
networkx==3.5
numpy==2.1.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.98.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
optree==0.14.1
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
patsy==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
phonenumbers==8.13.52
pillow==11.1.0
plotly==6.0.0
ply==3.11
posthog==5.4.0
preshed==3.0.9
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
pymongo==4.13.2
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-ulid==3.0.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
redisvl==0.8.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.13.0
SQLAlchemy==2.0.42
sqlparse==0.5.3
srsly==2.5.1
stanio==0.5.1
starlette==0.47.2
statsmodels==0.14.4
streamlit==1.44.1
sympy==1.14.0
tablib==3.5.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
toml==0.10.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.57.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
weasel==0.4.1
weasyprint==65.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.23.0
pydantic-settings
arxiv
pymongo
